Added SQS support and made library more robust.

+++ src/Commands/RunQueueServiceCommand.php
@@ -6,34 +6,51 @@ use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+use Symfony\Component\Console\Style\SymfonyStyle;
+use Symfony\Component\Process\Process;
+use Drutiny\Attribute\UsePlugin;
+use Drutiny\Plugin;
+
 
 #[AsCommand(
     name: 'bulk:run-queue-service',
     description: 'Run the queue service.'
 )]
+#[UsePlugin('queue:amqp')]
 class RunQueueServiceCommand extends Command
 {
-    use AMQPQueueTrait;
 
-    // ...
+    public function __construct(protected Plugin $plugin)
+    {
+        parent::__construct();
+    }
+
     protected function configure(): void
     {
         $this
-            // the command help shown when running the command with the "--help" option
+            ->setName('bulk:run-docker-rabbitmq')
+            ->setDescription('Runs a rabbitMQ server install a docker container.')
             ->setHelp('This command allows you to run the RabbitMQ queuing service inside a docker container.')
-            ->addQueueOptions()
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        // docker run --rm -it --hostname my-rabbit -p 15672:15672 -p 5672:5672 rabbitmq:3-management
+        $io = new SymfonyStyle($input, $output);
+        if ($this->plugin->host != 'localhost') {
+            $io->error("AMQP queue service is configured to connect to {$this->plugin->host} and would not connect to rabbitMQ server on localhost.");
+            return Command::INVALID;
+        }
+
         $command = strtr('docker run --rm -it --hostname drutiny-bulk-rabbitmq -p %admin_port%:%admin_port% -p %port%:%port% rabbitmq:3-management', [
-          '%admin_port%' => '1' . $input->getOption('ampq-port'),
-          '%port%' => $input->getOption('ampq-port'),
+          '%admin_port%' => '1' . $this->plugin->port,
+          '%port%' => $this->plugin->port,
         ]);
 
-        passthru($command, $exit_code);
-        return $exit_code;
+        $process = Process::fromShellCommandline($command);
+        $process->setTty(true);
+        $process->setTimeout(null);
+        $process->mustRun();
+        return $process->getExitCode();
     }
 }
